Fix deployed site being unable to find assets
It seems like the base URL I had was messing things up since it was set
to the GitHub Pages URL instead of the real domain. The actions for
setting up and configuring pages seems to already return the custom
domain now that it's configured, so I should be able to just always get
the correct domain and set the baseURL that way. Using
actions/configure-pages is shown in the current Hugo actions example, so
I'm guessing this is the fix that I need.