Add example images for SSH port forwarding
I really wanted to use svg images here which Excalidraw can export since
their file size is smaller and they can be resized really well, but for
some reason on the Hugo site the fonts look off. If I open the svg in a
browser it looks exactly as it did on excalidraw, but on the Hugo site
it uses a different font and the letters are off-center. For now I'll
just use pngs, but maybe I'll find a fix eventually.