Switch to Hugo w/ a custom theme
Honestly I'm just tired of the dependapot emails all the time for
Javascript dependencies and what not. I just want a simple static site,
and Hugo lets me do that.

When I used Hugo before, I always had some issues with the existing
themes for documentation-style sites. This time I opted to try and write
my own theme. This gives me the flexibility to change it as I want to,
and also allows me to keep it as simple as possible.

I've tested this site on desktop (chrome, firefox, safari) and safari on
iOS and it seems to be working everywhere. I'm sure there is some
styling I'm missing. I'm a trash frontend dev :).